Improved efficiency of computing point estimates
Previously we split on all unique combinations of the grouping column, and process each split separately. This is a valid approach, but can occur significant overhead when this forms a large number (>100k) groups. We now first determine which data (if any) require computation of a (bias-corrected) point estimate because of grouping.